% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{addEntrezID}
\alias{addEntrezID}
\title{FUNCTION_TITLE}
\usage{
addEntrezID(data)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
