% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{api_post_request}
\alias{api_post_request}
\title{FUNCTION_TITLE}
\usage{
api_post_request(route, params, retry_times, retry_pause_min)
}
\arguments{
\item{route}{PARAM_DESCRIPTION}

\item{params}{PARAM_DESCRIPTION}

\item{retry_times}{PARAM_DESCRIPTION}

\item{retry_pause_min}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}
\code{\link[httr]{add_headers}}, \code{\link[httr]{RETRY}}
\code{\link[jsonlite]{toJSON, fromJSON}}
}
