% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{api_request}
\alias{api_request}
\title{FUNCTION_TITLE}
\usage{
api_request(
  route,
  params,
  mode = c("table", "raw"),
  method = method,
  retry_times,
  retry_pause_min
)
}
\arguments{
\item{route}{PARAM_DESCRIPTION}

\item{params}{PARAM_DESCRIPTION}

\item{mode}{PARAM_DESCRIPTION, Default: c("table", "raw")}

\item{method}{PARAM_DESCRIPTION, Default: method}

\item{retry_times}{PARAM_DESCRIPTION}

\item{retry_pause_min}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[httr]{content}}
}
