% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{as.num}
\alias{as.num}
\title{FUNCTION_TITLE}
\usage{
as.num(x, na.strings = "NA")
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{na.strings}{PARAM_DESCRIPTION, Default: 'NA'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
