% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{cleanMem}
\alias{cleanMem}
\title{FUNCTION_TITLE}
\usage{
cleanMem(n = 10)
}
\arguments{
\item{n}{PARAM_DESCRIPTION, Default: 10}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
