% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{CleanNames}
\alias{CleanNames}
\title{FUNCTION_TITLE}
\usage{
CleanNames(query, type)
}
\arguments{
\item{query}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
