% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{CleanNumber}
\alias{CleanNumber}
\title{FUNCTION_TITLE}
\usage{
CleanNumber(bdata)
}
\arguments{
\item{bdata}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
