% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_feature.R
\name{coloc_plot}
\alias{coloc_plot}
\title{Title}
\usage{
coloc_plot(
  out = out,
  gwas1name = "gwas1",
  gwas2name = "gwas2",
  population = "EUR"
)
}
\arguments{
\item{gwas2name}{}
}
\description{
Title
}
