% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{color_scale}
\alias{color_scale}
\title{FUNCTION_TITLE}
\usage{
color_scale(c1 = "grey", c2 = "red")
}
\arguments{
\item{c1}{PARAM_DESCRIPTION, Default: 'grey'}

\item{c2}{PARAM_DESCRIPTION, Default: 'red'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
