% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetUtils.R
\name{Compute.SteinerForest}
\alias{Compute.SteinerForest}
\title{FUNCTION_TITLE}
\usage{
Compute.SteinerForest(ppi, terminals, w = 2, b = 1, mu = 5e-04, dummies)
}
\arguments{
\item{ppi}{PARAM_DESCRIPTION}

\item{terminals}{PARAM_DESCRIPTION}

\item{w}{PARAM_DESCRIPTION, Default: 2}

\item{b}{PARAM_DESCRIPTION, Default: 1}

\item{mu}{PARAM_DESCRIPTION, Default: 5e-04}

\item{dummies}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[igraph]{degree}}
}
