% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R, R/MiscUtils.R
\name{ComputeColorGradient}
\alias{ComputeColorGradient}
\title{FUNCTION_TITLE}
\usage{
ComputeColorGradient(nd.vec, background = "black", centered)

ComputeColorGradient(nd.vec, background = "black", centered)
}
\arguments{
\item{nd.vec}{PARAM_DESCRIPTION}

\item{background}{PARAM_DESCRIPTION, Default: 'black'}

\item{centered}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
