% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetUtils.R
\name{ComputeSubnetStats}
\alias{ComputeSubnetStats}
\title{FUNCTION_TITLE}
\usage{
ComputeSubnetStats(comps)
}
\arguments{
\item{comps}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
