% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetUtils.R
\name{convertIgraph2JSON}
\alias{convertIgraph2JSON}
\title{FUNCTION_TITLE}
\usage{
convertIgraph2JSON(g, filenm)
}
\arguments{
\item{g}{PARAM_DESCRIPTION}

\item{filenm}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[igraph]{edge_attr}}
}
