% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetUtils.R
\name{CreateGraph}
\alias{CreateGraph}
\title{FUNCTION_TITLE}
\usage{
CreateGraph(mSetObj = NA, net.type)
}
\arguments{
\item{mSetObj}{PARAM_DESCRIPTION, Default: NA}

\item{net.type}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
