% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{CrossReferencing}
\alias{CrossReferencing}
\title{FUNCTION_TITLE}
\usage{
CrossReferencing(
  mSetObj = NA,
  q.type,
  hmdb = T,
  pubchem = T,
  chebi = F,
  kegg = T,
  metlin = F,
  lipid = F
)
}
\arguments{
\item{mSetObj}{PARAM_DESCRIPTION, Default: NA}

\item{q.type}{PARAM_DESCRIPTION}

\item{hmdb}{PARAM_DESCRIPTION, Default: T}

\item{pubchem}{PARAM_DESCRIPTION, Default: T}

\item{chebi}{PARAM_DESCRIPTION, Default: F}

\item{kegg}{PARAM_DESCRIPTION, Default: T}

\item{metlin}{PARAM_DESCRIPTION, Default: F}

\item{lipid}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
