% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetUtils.R
\name{DecomposeGraph}
\alias{DecomposeGraph}
\title{FUNCTION_TITLE}
\usage{
DecomposeGraph(gObj, minNodeNum = 2)
}
\arguments{
\item{gObj}{PARAM_DESCRIPTION}

\item{minNodeNum}{PARAM_DESCRIPTION, Default: 2}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
