% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{doGeneIDMapping}
\alias{doGeneIDMapping}
\title{FUNCTION_TITLE}
\usage{
doGeneIDMapping(q.vec, org, type)
}
\arguments{
\item{q.vec}{PARAM_DESCRIPTION}

\item{org}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
