% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_feature.R
\name{download_ieuvcf}
\alias{download_ieuvcf}
\title{Batch download VCF files from ieu openGWAS database}
\usage{
download_ieuvcf(IDlist, path, timeout = 600)
}
\arguments{
\item{IDlist}{The GWAS ID in the ieu database you want to download}

\item{path}{The path where the file is saved}

\item{timeout}{Download timeout setting}
}
\description{
Batch download VCF files from ieu openGWAS database
}
\examples{
download_ieuvcf(c('bbj-a-86','bbj-a-94','bbj-a-99'), './', timeout = 600)
}
