% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetUtils.R
\name{ExcludeNodes}
\alias{ExcludeNodes}
\title{FUNCTION_TITLE}
\usage{
ExcludeNodes(nodeids, filenm)
}
\arguments{
\item{nodeids}{PARAM_DESCRIPTION}

\item{filenm}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
