% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{fast.write.csv}
\alias{fast.write.csv}
\title{FUNCTION_TITLE}
\usage{
fast.write.csv(dat, file, row.names = TRUE)
}
\arguments{
\item{dat}{PARAM_DESCRIPTION}

\item{file}{PARAM_DESCRIPTION}

\item{row.names}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[data.table]{fwrite}}
}
