% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetUtils.R
\name{FilterBipartiNet}
\alias{FilterBipartiNet}
\title{FUNCTION_TITLE}
\usage{
FilterBipartiNet(mSetObj = NA, nd.type, min.dgr, min.btw)
}
\arguments{
\item{mSetObj}{PARAM_DESCRIPTION, Default: NA}

\item{nd.type}{PARAM_DESCRIPTION}

\item{min.dgr}{PARAM_DESCRIPTION}

\item{min.btw}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
