% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetUtils.R
\name{FilterNetByList}
\alias{FilterNetByList}
\title{FUNCTION_TITLE}
\usage{
FilterNetByList(net.type, ids, id.type, remove)
}
\arguments{
\item{net.type}{PARAM_DESCRIPTION}

\item{ids}{PARAM_DESCRIPTION}

\item{id.type}{PARAM_DESCRIPTION}

\item{remove}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
