% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetUtils.R
\name{FindCommunities}
\alias{FindCommunities}
\title{FUNCTION_TITLE}
\usage{
FindCommunities(method = "walktrap", use.weight = FALSE)
}
\arguments{
\item{method}{PARAM_DESCRIPTION, Default: 'walktrap'}

\item{use.weight}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
