% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{flatten_response}
\alias{flatten_response}
\title{FUNCTION_TITLE}
\usage{
flatten_response(response, field = "results")
}
\arguments{
\item{response}{PARAM_DESCRIPTION}

\item{field}{PARAM_DESCRIPTION, Default: 'results'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[httr]{content}}
\code{\link[rjson]{fromJSON}}
\code{\link[purrr]{pluck}}
\code{\link[tibble]{as_tibble}}
}
