% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R, R/MiscUtils.R
\name{generate_breaks}
\alias{generate_breaks}
\title{FUNCTION_TITLE}
\usage{
generate_breaks(x, n, center = F)

generate_breaks(x, n, center = F)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{n}{PARAM_DESCRIPTION}

\item{center}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
