% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{get_ld_proxies}
\alias{get_ld_proxies}
\title{FUNCTION_TITLE}
\usage{
get_ld_proxies(
  rsid,
  bfile,
  searchspace = NULL,
  tag_kb = 5000,
  tag_nsnp = 5000,
  tag_r2 = 0.6,
  threads = 1,
  out = paste0(getwd(), "/")
)
}
\arguments{
\item{rsid}{PARAM_DESCRIPTION}

\item{bfile}{PARAM_DESCRIPTION}

\item{searchspace}{PARAM_DESCRIPTION, Default: NULL}

\item{tag_kb}{PARAM_DESCRIPTION, Default: 5000}

\item{tag_nsnp}{PARAM_DESCRIPTION, Default: 5000}

\item{tag_r2}{PARAM_DESCRIPTION, Default: 0.6}

\item{threads}{PARAM_DESCRIPTION, Default: 1}

\item{out}{PARAM_DESCRIPTION, Default: paste0(getwd(), "/")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[gwasvcf]{check_plink}}
\code{\link[utils]{write.table}}
\code{\link[data.table]{fread}}
\code{\link[dplyr]{reexports}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{desc}}
}
