% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{getApiResult}
\alias{getApiResult}
\title{FUNCTION_TITLE}
\usage{
getApiResult(url = "NA", init = TRUE)
}
\arguments{
\item{url}{PARAM_DESCRIPTION, Default: 'NA'}

\item{init}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rjson]{fromJSON}}
}
