% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{GetCompletePheMRMetDisList}
\alias{GetCompletePheMRMetDisList}
\title{FUNCTION_TITLE}
\usage{
GetCompletePheMRMetDisList()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
