% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{getDataFromTextArea}
\alias{getDataFromTextArea}
\title{FUNCTION_TITLE}
\usage{
getDataFromTextArea(txtInput, sep.type = "space")
}
\arguments{
\item{txtInput}{PARAM_DESCRIPTION}

\item{sep.type}{PARAM_DESCRIPTION, Default: 'space'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
