% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{getHaploregData}
\alias{getHaploregData}
\title{FUNCTION_TITLE}
\usage{
getHaploregData(url, body, timeout)
}
\arguments{
\item{url}{PARAM_DESCRIPTION}

\item{body}{PARAM_DESCRIPTION}

\item{timeout}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
