% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetUtils.R
\name{GetMinConnectedGraphs}
\alias{GetMinConnectedGraphs}
\title{FUNCTION_TITLE}
\usage{
GetMinConnectedGraphs(net.type, max.len = 200)
}
\arguments{
\item{net.type}{PARAM_DESCRIPTION}

\item{max.len}{PARAM_DESCRIPTION, Default: 200}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
