% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunUtils.R
\name{GetMRMatRowNames}
\alias{GetMRMatRowNames}
\title{FUNCTION_TITLE}
\usage{
GetMRMatRowNames(mSetObj = NA, type)
}
\arguments{
\item{mSetObj}{PARAM_DESCRIPTION, Default: NA}

\item{type}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
