% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunUtils.R
\name{GetMRRes.rowNames}
\alias{GetMRRes.rowNames}
\title{FUNCTION_TITLE}
\usage{
GetMRRes.rowNames(mSetObj = NA)
}
\arguments{
\item{mSetObj}{PARAM_DESCRIPTION, Default: NA}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
