% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetUtils.R
\name{GetNetStatByType}
\alias{GetNetStatByType}
\title{FUNCTION_TITLE}
\usage{
GetNetStatByType(g)
}
\arguments{
\item{g}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
