% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetUtils.R
\name{GetNetStats}
\alias{GetNetStats}
\title{FUNCTION_TITLE}
\usage{
GetNetStats()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
