% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{GetResCol}
\alias{GetResCol}
\title{FUNCTION_TITLE}
\usage{
GetResCol(netType, colInx)
}
\arguments{
\item{netType}{PARAM_DESCRIPTION}

\item{colInx}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
