% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{GetResRowNames}
\alias{GetResRowNames}
\title{FUNCTION_TITLE}
\usage{
GetResRowNames(netType)
}
\arguments{
\item{netType}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
