% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetUtils.R
\name{GetShortestPaths}
\alias{GetShortestPaths}
\title{FUNCTION_TITLE}
\usage{
GetShortestPaths(from, to)
}
\arguments{
\item{from}{PARAM_DESCRIPTION}

\item{to}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
