% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_feature.R
\name{ieu_finemap}
\alias{ieu_finemap}
\title{Title}
\usage{
ieu_finemap(
  region,
  id,
  bfile,
  plink_bin = genetics.binaRies::get_plink_binary()
)
}
\arguments{
\item{plink_bin}{}
}
\description{
Title
}
