% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{InitDataObjects}
\alias{InitDataObjects}
\title{FUNCTION_TITLE}
\usage{
InitDataObjects(anal.type)
}
\arguments{
\item{anal.type}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
