% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{MetaboliteMappingExact}
\alias{MetaboliteMappingExact}
\title{FUNCTION_TITLE}
\usage{
MetaboliteMappingExact(mSetObj = NA, q.type, lipid = F)
}
\arguments{
\item{mSetObj}{PARAM_DESCRIPTION, Default: NA}

\item{q.type}{PARAM_DESCRIPTION}

\item{lipid}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
