% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mGWASR-package.R
\docType{package}
\name{mGWASR-package}
\alias{mGWASR}
\alias{mGWASR-package}
\title{mGWASR: A companion R package for the mGWAS-Explorer web server}
\description{
This package contains the entire underlying R functionality of the web-based tool mGWAS-Explorer.
}
\author{
\strong{Maintainer}: Le Chang \email{le.chang@mail.mcgill.ca} (\href{https://orcid.org/0000-0002-0966-7923}{ORCID})

Authors:
\itemize{
  \item Jianguo Xia \email{jeff.xia@mcgill.ca} (\href{https://orcid.org/0000-0003-2040-2624}{ORCID})
}

}
\keyword{internal}
