% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_feature.R
\name{pair_ldsc}
\alias{pair_ldsc}
\title{Title}
\usage{
pair_ldsc(
  path,
  output_dir,
  ancestry = "EUR",
  para_plan = "multicore",
  num_cores = 3,
  log_file = "ldsc.log"
)
}
\arguments{
\item{log_file}{}
}
\description{
Title
}
