% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{PerformCmpdMapping}
\alias{PerformCmpdMapping}
\title{FUNCTION_TITLE}
\usage{
PerformCmpdMapping(mSetObj = NA, cmpdIDs, idType, tissueType, population)
}
\arguments{
\item{mSetObj}{PARAM_DESCRIPTION, Default: NA}

\item{cmpdIDs}{PARAM_DESCRIPTION}

\item{idType}{PARAM_DESCRIPTION}

\item{tissueType}{PARAM_DESCRIPTION}

\item{population}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
