% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunUtils.R
\name{PerformEnrichAnalysis}
\alias{PerformEnrichAnalysis}
\title{FUNCTION_TITLE}
\usage{
PerformEnrichAnalysis(file.nm, fun.type, ora.vec)
}
\arguments{
\item{file.nm}{PARAM_DESCRIPTION}

\item{fun.type}{PARAM_DESCRIPTION}

\item{ora.vec}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[qs]{qread}}
\code{\link[rjson]{toJSON}}
}
