% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetUtils.R
\name{PerformLayOut}
\alias{PerformLayOut}
\title{FUNCTION_TITLE}
\usage{
PerformLayOut(g, layers, algo, focus = "")
}
\arguments{
\item{g}{PARAM_DESCRIPTION}

\item{layers}{PARAM_DESCRIPTION}

\item{algo}{PARAM_DESCRIPTION}

\item{focus}{PARAM_DESCRIPTION, Default: ''}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
