% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunUtils.R
\name{PerformMRAnalysis}
\alias{PerformMRAnalysis}
\title{FUNCTION_TITLE}
\usage{
PerformMRAnalysis(
  ldclumpOpt,
  ldProxies,
  ldThresh,
  pldSNPs,
  mafThresh,
  harmonizeOpt
)
}
\arguments{
\item{ldclumpOpt}{PARAM_DESCRIPTION}

\item{ldProxies}{PARAM_DESCRIPTION}

\item{ldThresh}{PARAM_DESCRIPTION}

\item{pldSNPs}{PARAM_DESCRIPTION}

\item{mafThresh}{PARAM_DESCRIPTION}

\item{harmonizeOpt}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[TwoSampleMR]{clump_data}}, \code{\link[TwoSampleMR]{extract_outcome_data}}, \code{\link[TwoSampleMR]{harmonise_data}}, \code{\link[TwoSampleMR]{mr}}, \code{\link[TwoSampleMR]{mr_heterogeneity}}, \code{\link[TwoSampleMR]{mr_pleiotropy_test}}, \code{\link[TwoSampleMR]{mr_singlesnp}}, \code{\link[TwoSampleMR]{mr_leaveoneout}}
}
