% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunUtils.R
\name{PerformNetEnrichment}
\alias{PerformNetEnrichment}
\title{FUNCTION_TITLE}
\usage{
PerformNetEnrichment(file.nm, fun.type, IDs)
}
\arguments{
\item{file.nm}{PARAM_DESCRIPTION}

\item{fun.type}{PARAM_DESCRIPTION}

\item{IDs}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
