% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{phenoscanner}
\alias{phenoscanner}
\title{FUNCTION_TITLE}
\usage{
phenoscanner(
  snpquery = NULL,
  genequery = NULL,
  regionquery = NULL,
  catalogue = "GWAS",
  pvalue = 1e-05,
  proxies = "None",
  r2 = 0.8,
  build = 37
)
}
\arguments{
\item{snpquery}{PARAM_DESCRIPTION, Default: NULL}

\item{genequery}{PARAM_DESCRIPTION, Default: NULL}

\item{regionquery}{PARAM_DESCRIPTION, Default: NULL}

\item{catalogue}{PARAM_DESCRIPTION, Default: 'GWAS'}

\item{pvalue}{PARAM_DESCRIPTION, Default: 1e-05}

\item{proxies}{PARAM_DESCRIPTION, Default: 'None'}

\item{r2}{PARAM_DESCRIPTION, Default: 0.8}

\item{build}{PARAM_DESCRIPTION, Default: 37}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
