% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetUtils.R
\name{PlotDegreeHistogram}
\alias{PlotDegreeHistogram}
\title{FUNCTION_TITLE}
\usage{
PlotDegreeHistogram(imgNm, netNm = "NA", dpi = 72, format = "png")
}
\arguments{
\item{imgNm}{PARAM_DESCRIPTION}

\item{netNm}{PARAM_DESCRIPTION, Default: 'NA'}

\item{dpi}{PARAM_DESCRIPTION, Default: 72}

\item{format}{PARAM_DESCRIPTION, Default: 'png'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
