% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRGraphics.R
\name{PlotLeaveOneOut}
\alias{PlotLeaveOneOut}
\title{FUNCTION_TITLE}
\usage{
PlotLeaveOneOut(mSetObj = NA, imgName, format = "png", dpi = 72, width = NA)
}
\arguments{
\item{mSetObj}{PARAM_DESCRIPTION, Default: NA}

\item{imgName}{PARAM_DESCRIPTION}

\item{format}{PARAM_DESCRIPTION, Default: 'png'}

\item{dpi}{PARAM_DESCRIPTION, Default: 72}

\item{width}{PARAM_DESCRIPTION, Default: NA}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[Cairo]{Cairo}}
}
