% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetUtils.R
\name{PrepareGraphML}
\alias{PrepareGraphML}
\title{FUNCTION_TITLE}
\usage{
PrepareGraphML(net.nm)
}
\arguments{
\item{net.nm}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
