% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{Query.mGWASDB}
\alias{Query.mGWASDB}
\title{FUNCTION_TITLE}
\usage{
Query.mGWASDB(
  db.path,
  q.vec,
  table.nm,
  col.nm,
  biofluid = "all",
  population = "all",
  db.opt = "kegg"
)
}
\arguments{
\item{db.path}{PARAM_DESCRIPTION}

\item{q.vec}{PARAM_DESCRIPTION}

\item{table.nm}{PARAM_DESCRIPTION}

\item{col.nm}{PARAM_DESCRIPTION}

\item{biofluid}{PARAM_DESCRIPTION, Default: 'all'}

\item{population}{PARAM_DESCRIPTION, Default: 'all'}

\item{db.opt}{PARAM_DESCRIPTION, Default: 'kegg'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
