% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{Query.pQTLDB}
\alias{Query.pQTLDB}
\title{FUNCTION_TITLE}
\usage{
Query.pQTLDB(db.path, q.mat)
}
\arguments{
\item{db.path}{PARAM_DESCRIPTION}

\item{q.mat}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
