% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{query_melodipresto}
\alias{query_melodipresto}
\title{FUNCTION_TITLE}
\usage{
query_melodipresto(
  route,
  params = NULL,
  mode = c("raw", "table"),
  method = c("GET", "POST"),
  retry_times = 3,
  retry_pause_min = 1
)
}
\arguments{
\item{route}{PARAM_DESCRIPTION}

\item{params}{PARAM_DESCRIPTION, Default: NULL}

\item{mode}{PARAM_DESCRIPTION, Default: c("raw", "table")}

\item{method}{PARAM_DESCRIPTION, Default: c("GET", "POST")}

\item{retry_times}{PARAM_DESCRIPTION, Default: 3}

\item{retry_pause_min}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{as_tibble}}
}
