% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunUtils.R
\name{QueryLiteratureMelodiPresto}
\alias{QueryLiteratureMelodiPresto}
\title{FUNCTION_TITLE}
\usage{
QueryLiteratureMelodiPresto(exposure, outcome)
}
\arguments{
\item{exposure}{PARAM_DESCRIPTION}

\item{outcome}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
