% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtils.R
\name{QueryMultiList}
\alias{QueryMultiList}
\title{FUNCTION_TITLE}
\usage{
QueryMultiList()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
