% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{QueryMyVariant}
\alias{QueryMyVariant}
\title{FUNCTION_TITLE}
\usage{
QueryMyVariant(query, scope)
}
\arguments{
\item{query}{PARAM_DESCRIPTION}

\item{scope}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[myvariant]{getVariants}}
\code{\link[mygene]{queryMany}}
}
