% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtils.R
\name{QueryOutcome}
\alias{QueryOutcome}
\title{FUNCTION_TITLE}
\usage{
QueryOutcome(itemVec)
}
\arguments{
\item{itemVec}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
