% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtils.R
\name{QueryPheMR}
\alias{QueryPheMR}
\title{FUNCTION_TITLE}
\usage{
QueryPheMR(itemVec, resOpt)
}
\arguments{
\item{itemVec}{PARAM_DESCRIPTION}

\item{resOpt}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
