% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtils.R
\name{QuerySigPheMR}
\alias{QuerySigPheMR}
\title{FUNCTION_TITLE}
\usage{
QuerySigPheMR()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[qs]{qread}}
}
