% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtils.R
\name{QuerySingleItem}
\alias{QuerySingleItem}
\title{FUNCTION_TITLE}
\usage{
QuerySingleItem(idType, itemVec)
}
\arguments{
\item{idType}{PARAM_DESCRIPTION}

\item{itemVec}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
