% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtils.R
\name{QueryStudy}
\alias{QueryStudy}
\title{FUNCTION_TITLE}
\usage{
QueryStudy(pmid)
}
\arguments{
\item{pmid}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
