% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{QueryVEP}
\alias{QueryVEP}
\title{FUNCTION_TITLE}
\usage{
QueryVEP(q.vec, vepDis, content_type = "application/json")
}
\arguments{
\item{q.vec}{PARAM_DESCRIPTION}

\item{vepDis}{PARAM_DESCRIPTION}

\item{content_type}{PARAM_DESCRIPTION, Default: 'application/json'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
