% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetUtils.R
\name{ReduceEdgeDensity}
\alias{ReduceEdgeDensity}
\title{FUNCTION_TITLE}
\usage{
ReduceEdgeDensity(nd.type = "all")
}
\arguments{
\item{nd.type}{PARAM_DESCRIPTION, Default: 'all'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
