% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{RemoveNode}
\alias{RemoveNode}
\title{FUNCTION_TITLE}
\usage{
RemoveNode(node.id)
}
\arguments{
\item{node.id}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
