% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{rescale}
\alias{rescale}
\title{FUNCTION_TITLE}
\usage{
rescale(x, from, to)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{from}{PARAM_DESCRIPTION}

\item{to}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
