% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{rescale2NewRange}
\alias{rescale2NewRange}
\title{FUNCTION_TITLE}
\usage{
rescale2NewRange(qvec, a, b)
}
\arguments{
\item{qvec}{PARAM_DESCRIPTION}

\item{a}{PARAM_DESCRIPTION}

\item{b}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
