% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R, R/MiscUtils.R
\name{scale_vec_colours}
\alias{scale_vec_colours}
\title{FUNCTION_TITLE}
\usage{
scale_vec_colours(x, col = rainbow(10), breaks = NA)

scale_vec_colours(x, col = rainbow(10), breaks = NA)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{col}{PARAM_DESCRIPTION, Default: rainbow(10)}

\item{breaks}{PARAM_DESCRIPTION, Default: NA}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
