% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtils.R
\name{SearchMultiNet}
\alias{SearchMultiNet}
\title{FUNCTION_TITLE}
\usage{
SearchMultiNet(input.type)
}
\arguments{
\item{input.type}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
