% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{SetAnalType}
\alias{SetAnalType}
\title{FUNCTION_TITLE}
\usage{
SetAnalType(analType)
}
\arguments{
\item{analType}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
