% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtils.R
\name{SetPpiDb}
\alias{SetPpiDb}
\title{FUNCTION_TITLE}
\usage{
SetPpiDb(db, req, conf)
}
\arguments{
\item{db}{PARAM_DESCRIPTION}

\item{req}{PARAM_DESCRIPTION}

\item{conf}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
