% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{SetupIndListData}
\alias{SetupIndListData}
\title{FUNCTION_TITLE}
\usage{
SetupIndListData(mSetObj = NA, mirs, inputType, idType, tissue, population)
}
\arguments{
\item{mSetObj}{PARAM_DESCRIPTION, Default: NA}

\item{mirs}{PARAM_DESCRIPTION}

\item{inputType}{PARAM_DESCRIPTION}

\item{idType}{PARAM_DESCRIPTION}

\item{tissue}{PARAM_DESCRIPTION}

\item{population}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
