% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtils.R
\name{SetVepOpt}
\alias{SetVepOpt}
\title{FUNCTION_TITLE}
\usage{
SetVepOpt(opt)
}
\arguments{
\item{opt}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
