% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{ShowMemoryUse}
\alias{ShowMemoryUse}
\title{FUNCTION_TITLE}
\usage{
ShowMemoryUse(..., n = 40)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}

\item{n}{PARAM_DESCRIPTION, Default: 40}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
