% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{simpleQuery}
\alias{simpleQuery}
\title{FUNCTION_TITLE}
\usage{
simpleQuery(
  query = NULL,
  file = NULL,
  study = NULL,
  ldThresh = 0.8,
  ldPop = "EUR",
  epi = "vanilla",
  cons = "siphy",
  genetypes = "gencode",
  url = Haploreg.settings[["base.url"]],
  timeout = 100,
  encoding = "UTF-8",
  querySNP = FALSE,
  fields = NULL,
  verbose = FALSE
)
}
\arguments{
\item{query}{PARAM_DESCRIPTION, Default: NULL}

\item{file}{PARAM_DESCRIPTION, Default: NULL}

\item{study}{PARAM_DESCRIPTION, Default: NULL}

\item{ldThresh}{PARAM_DESCRIPTION, Default: 0.8}

\item{ldPop}{PARAM_DESCRIPTION, Default: 'EUR'}

\item{epi}{PARAM_DESCRIPTION, Default: 'vanilla'}

\item{cons}{PARAM_DESCRIPTION, Default: 'siphy'}

\item{genetypes}{PARAM_DESCRIPTION, Default: 'gencode'}

\item{url}{PARAM_DESCRIPTION, Default: Haploreg.settings[\link{"base.url"}]}

\item{timeout}{PARAM_DESCRIPTION, Default: 100}

\item{encoding}{PARAM_DESCRIPTION, Default: 'UTF-8'}

\item{querySNP}{PARAM_DESCRIPTION, Default: FALSE}

\item{fields}{PARAM_DESCRIPTION, Default: NULL}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
