% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{stop_for_status}
\alias{stop_for_status}
\title{FUNCTION_TITLE}
\usage{
stop_for_status(response, context)
}
\arguments{
\item{response}{PARAM_DESCRIPTION}

\item{context}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[httr]{status_code}}, \code{\link[httr]{http_condition}}
}
