% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{UpdateEntries}
\alias{UpdateEntries}
\title{FUNCTION_TITLE}
\usage{
UpdateEntries(mSetObj = NA, col.id, method, value, action)
}
\arguments{
\item{mSetObj}{PARAM_DESCRIPTION, Default: NA}

\item{col.id}{PARAM_DESCRIPTION}

\item{method}{PARAM_DESCRIPTION}

\item{value}{PARAM_DESCRIPTION}

\item{action}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
