% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetUtils.R
\name{UpdateNetworkLayout}
\alias{UpdateNetworkLayout}
\title{FUNCTION_TITLE}
\usage{
UpdateNetworkLayout(algo, filenm, focus)
}
\arguments{
\item{algo}{PARAM_DESCRIPTION}

\item{filenm}{PARAM_DESCRIPTION}

\item{focus}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
