% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_feature.R
\name{vcf_finemap}
\alias{vcf_finemap}
\title{Title}
\usage{
vcf_finemap(
  region,
  vcf,
  bfile,
  plink_bin = genetics.binaRies::get_plink_binary(),
  threads = 1
)
}
\arguments{
\item{threads}{}
}
\description{
Title
}
